requirements.txt
anthropic>=0.40.0
apify-client>=1.8.0
requests>=2.32.0
python-dotenv>=1.0.0
